不要なノードを削除した

+++ App/BhDefine/EventHandler/Exp/OnChildSoundExpReplaced.js
@@ -6,30 +6,34 @@
 
   let nextMelodyExp = bhThis.findDescendantOf('*', 'Arg1', '*');
 
+  // 音削除
+  if (String(bhReplacedNewNode.getSymbolName()) === 'SoundExpVoid') {
+    if (String(nextMelodyExp.getSymbolName()) === 'MelodyExp') {
+      bhNodePlacer.moveToWs(bhThis.getWorkspace(), nextMelodyExp, 0, 0, bhUserOpe);
+      bhNodePlacer.replaceChild(bhThis, nextMelodyExp, bhUserOpe);
+      bhNodePlacer.deleteNode(bhThis, bhUserOpe);
+    }
+
   // 音挿入
-  if (String(newNodeSectionName) === 'SoundExp') {
+  } else if (String(newNodeSectionName) === 'SoundExp') {
 
     // 新メロディノード作成
     let posOnWS = bhUtil.getPosOnWorkspace(bhThis) ?? {x: 0, y: 0};
     let newMelodyExp = bhUtil.addNewNodeToWS(
-      'idMelodyExp', bhThis.getWorkspace(), posOnWS, bhUserOpe);
+        'idMelodyExp', bhThis.getWorkspace(), posOnWS, bhUserOpe);
     bhNodePlacer.exchangeNodes(nextMelodyExp, newMelodyExp, bhUserOpe);
-    
+
     // 新ノードの末尾に旧ノードを追加
     let newMelodyExpNext = newMelodyExp.findDescendantOf('*', 'Arg1', '*');
     bhNodePlacer.exchangeNodes(newMelodyExpNext, nextMelodyExp, bhUserOpe);
     bhNodePlacer.deleteNode(newMelodyExpNext, bhUserOpe);
-    
-    // 旧音ノードを新メロディノードに繋ぎ直し
-    let soundExpVoid = newMelodyExp.findDescendantOf('*', 'Arg0', '*');
-    bhNodePlacer.exchangeNodes(soundExpVoid, bhReplacedOldNode, bhUserOpe);
-    bhNodePlacer.deleteNode(soundExpVoid, bhUserOpe);
-  }
-  // 音削除
-  else if (String(bhParentConnector.getConnectedNode().getSymbolName()) === 'SoundExpVoid') {
-    if (String(nextMelodyExp.getSymbolName()) === 'MelodyExp') {
-      bhNodePlacer.replaceChild(bhThis, nextMelodyExp, bhUserOpe);
-      bhNodePlacer.deleteNode(bhThis, bhUserOpe);
+
+    // bhReplacedOldNode が SoundExpVoid の場合この時点で削除されている.
+    if (!bhReplacedOldNode.isDeleted()) {
+      // 旧音ノードを新メロディノードに繋ぎ直し
+      let soundExpVoid = newMelodyExp.findDescendantOf('*', 'Arg0', '*');
+      bhNodePlacer.exchangeNodes(soundExpVoid, bhReplacedOldNode, bhUserOpe);
+      bhNodePlacer.deleteNode(soundExpVoid, bhUserOpe);
     }
   }
 })();
